Multicast moved to Individual Server component
Move to a more solid implementation but not yet all the way. Server
listens and adds Sequence number and SOFH header.
But there is no redundancy yet, if this component fails or if there is a
problem with the UDP stream then the data is lost. Of course in a
real-world implementation this would be resolved through use of A-feed,
B-feed, C-feed model